video 2 terminado

+++ backEnd/controllers/assetController.js
@@ -3,9 +3,10 @@
 const asyncHandler = require("express-async-handler");
 const { get, request } = require("http");
 const Asset = require("../model/AssetModel");
+const User = require("../model/userModel");
 
 const getAsset = asyncHandler( async(req, res) =>{
-    const asset = await Asset.find({});
+    const asset = await Asset.find({ user: req.user.id });
     res.status(200).json(asset);
 });
 
@@ -14,7 +15,8 @@ const postAsset = asyncHandler (async (req, res, next) => {
         return res.status(400).json({ message: "El campo de texto del asset es requerido" });
     }
     const asset = await Asset.create({
-        text : req.body.text
+        text : req.body.text,
+        user: req.user.id
     });
     res.status(200).json(asset);
 
@@ -27,12 +29,35 @@ const putAsset = asyncHandler (async(req, res) =>{
     if(!asset){
         throw new Error('Asset no encontrado');
     }
+    const user = await User.findById(req.user.id);
+    
+    if(!user){
+        res.status(404);
+        throw new Error('Usuario no encontrado');
+    }
+
+    if(asset.user.toString() !== user.id){
+        res.status(401);
+        throw new Error('No autorizado');
+    }
 
     const updatedAsset = await Asset.findByIdAndUpdate(req.params.id,req.body,{new:true})
     res.status(200).json(updatedAsset);
 });
 
 const deleteAsset = asyncHandler (async(req, res) =>{
+    const user = await User.findById(req.user.id);
+    
+    if(!user){
+        res.status(404);
+        throw new Error('Usuario no encontrado');
+    }
+
+    if(asset.user.toString() !== user.id){
+        res.status(401);
+        throw new Error('No autorizado');
+    }
+    
     const asset = await Asset.deleteOne({ _id: req.params.id })
     res.status(200).json(asset);
 });
